% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Angola_TES_genotype_matrix}
\alias{Angola_TES_genotype_matrix}
\title{Genotypic data from a TES conducted in Angola by Dimbu et al (2024)}
\format{
A named list of binary (0/1) genotype matrices: list names correspond
to markers; matrix row names correspond to isolates; matrix column names
correspond to alleles; and matrix elements are set to 1 if the relevant allele
is called in a given isolate at that marker, and zero otherwise.
}
\source{
\url{https://github.com/MateuszPlucinski/AngolaTES2021}
}
\usage{
Angola_TES_genotype_matrix
}
\description{
This contains genotypes at a panel of 7 neutral microsatellite markers
(M313, M383, TA1, POLYA, PFPK2, M2490, TA109) for 182 isolates sampled
across 3 study sites (Benguela, Lunda Sul and Zaire) in a TES conducted
in Angola by Dimbu et al (2024)
}
\details{
Dimbu PR, Labuda S, Ferreira CM, Caquece F, André K, Pembele G, Pode D, João MF,
Pelenda VM, Nieto Andrade B, Horton B. Therapeutic response to four artemisinin-based
combination therapies in Angola, 2021. Antimicrobial Agents and Chemotherapy.
2024 Apr 3;68(4):e01525-23.
}
\keyword{datasets}
