% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{convolve_over_detected_clones}
\alias{convolve_over_detected_clones}
\title{Convolve the conditional likelihood over the distribution of detected clones}
\usage{
convolve_over_detected_clones(
  omega,
  cond_lhood,
  recurrent_locus_card,
  recurrent_MOI,
  ref_C_MOI
)
}
\arguments{
\item{omega}{Locus-wise per-clone probability of detection (applied to the
paired baseline and recurrent isolates only)}

\item{cond_lhood}{Data frame of conditional likelihoods generated with the
function locus_likelihood_given_detected_clones: this encodes the conditional
locus-wise likelihood (p) of the observed genotypes for the recurrent isolate,
given the number of newly-inoculated (n_I) and recrudescent (n_C) clones
comprising the recurrent isolate; and the number of succesfully genotyped clones
in the paired baseline (genotyped_C) and recurrent (genotyped_recurrent) isolates
respectively}

\item{recurrent_locus_card}{cardinality of the recurrent locus at the given marker}

\item{recurrent_MOI}{MOI of the recurrent isolate}

\item{ref_C_MOI}{MOI for the paired baseline isolate}
}
\value{
A data frame for the locus-wise likelihood (lhood) of the observed
genotypes for the recurrent isolate, stratified by the number of newly-inoculated (n_I)
vs recrudescent (n_C) clones
}
\description{
This function evaluates convolves the conditional locus-wise likelihood of
observations for the recurrent isolate, given the number of successfully genotyped
clones in the paired baseline and recurrent isolates, over the distribution
of detected clones
}
