% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_moments.R
\name{dirichlet_transform}
\alias{dirichlet_transform}
\title{Convert moments of sample allele counts to moments of population allele
frequencies derived under a multinomial-Dirichlet model (with a uniform
prior over the standard simplex)}
\usage{
dirichlet_transform(
  pop_moments_not_in_subset,
  recurrent_MOI,
  n_recurrent_alleles,
  n_possible_alleles,
  n_clones
)
}
\arguments{
\item{pop_moments_not_in_subset}{A matrix of moments of sample allele counts:
columns correspond to allelic subsets (governed by inclusion_exclusion_mask),
rows correspond to the moment ordinal (starting from zero)}

\item{recurrent_MOI}{Maximum moment ordinal}

\item{n_recurrent_alleles}{Number of distinct alleles over which allele counts
have been summed (the columns of pop_moments_not_in_subset iterate over the
relevant power set)}

\item{n_possible_alleles}{Number of possible alleles at the relevant locus}

\item{n_clones}{Number of clones over which allele counts have been computed}
}
\value{
A matrix of moments of moments of sample allele frequencies
}
\description{
Convert moments of sample allele counts to moments of population allele
frequencies derived under a multinomial-Dirichlet model (with a uniform
prior over the standard simplex)
}
