% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_posterior.R
\name{evaluate_posterior}
\alias{evaluate_posterior}
\title{Posterior summary}
\usage{
evaluate_posterior(
  recurrent_isolate,
  ref_C,
  ref_I,
  genotype_matrix,
  error_matrix,
  omega_vals = seq(0.75, 1, 0.05),
  keep_markers = "all"
)
}
\arguments{
\item{recurrent_isolate}{Name (string) of the recurrent isolate for which classification
is to be performed}

\item{ref_C}{Name (string) of the baseline isolate paired to the recurrent isolate}

\item{ref_I}{Vector of names (strings) of baseline isolates that are NOT
paired to the recurrent isolate, but used to estimate allele frequencies for
newly-inoculated clones}

\item{genotype_matrix}{Named marker-wise list where each element is a binary (0/1)
matrix of genotypes at a given marker: named rows correspond to isolates; named columns
correspond to alleles; and matrix elements are set to 1 if the relevant allele
is called in a given isolate, and zero otherwise. Every row must have sum at
least one; isolates with no alleles called at a given marker are dropped from the
genotype matrix for the relevant marker.}

\item{error_matrix}{Named marker-wise list where each element is a row stochastic
matrix (with row and column names corresponding to alleles at a given marker),
where element (i,j) yields the probability that an allele called as i in a
baseline isolate matches an allele called as j in the recurrent isolate at that marker}

\item{omega_vals}{Vector of values for omega, governing the locus-wise per-clone
probability of detection (applied to the paired baseline and recurrent isolates only)}

\item{keep_markers}{Set to "all" (by default) to calculate the posterior over
all markers; otherwise, a vector of marrkers (strings) over which to calculate
the posterior}
}
\value{
A list of two elements: element "posterior" is a data frame for the
posterior distribution (posterior) of newly-inoculated (n_I) vs recrudescent (n_C)
clones in the recurrent isolate, stratified by the per-clone probability of
detection for the paired baseline/recurrent isolates (omega); element "metrics"
is a data frame detailing the posterior probability of at least one recrudescent
clone (M1) and the expected proprtion of recrudescent clones (M2) stratified by omega
}
\description{
This function computes the posterior distribution of recrudescent vs
newly-inoculated clones within a recurrent isolate
}
