% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_matrices.R
\name{geometric_error_matrix}
\alias{geometric_error_matrix}
\title{Normalised geometric genotyping error model}
\usage{
geometric_error_matrix(marker_set, EPSILON)
}
\arguments{
\item{marker_set}{A named marker-wise list of vectors, listing all possible
alleles at each marker (must be strings, but encoding integers)}

\item{EPSILON}{Genotyping error rate}
}
\value{
A named marker-wise list of row stochastic matrices (with row and column names
corresponding to alleles), where element (i,j) yields the probability
that an allele called as i in a baseline isolate matches an allele called
as j in the recurrent isolate
}
\description{
Length-dependent genotyping error, adapted to length polymorphic markers with
integer lengths: genotyping error rate EPSILON, with a geometric dependence
on the allele length difference
}
