% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{locus_likelihood}
\alias{locus_likelihood}
\title{Locuswise likelihood}
\usage{
locus_likelihood(
  recurrent_alleles,
  recurrent_MOI,
  ref_I_genotypes,
  ref_I_MOI,
  ref_C_genotypes,
  ref_C_MOI,
  error_matrix,
  omega_vals
)
}
\arguments{
\item{recurrent_alleles}{Vector of alleles (strings) observed in the recurrent isolate}

\item{recurrent_MOI}{MOI of the recurrent isolate}

\item{ref_I_genotypes}{A binary (0/1) matrix of genotypes for baseline isolates
over which population I allele frequencies are derived: named rows correspond to isolates;
named columns correspond to alleles; and matrix elements are set to 1 if the
relevant allele is called in a given isolate, and zero otherwise}

\item{ref_I_MOI}{A named vector of MOIs for baseline isolates over which
population I allele frequencies are derived}

\item{ref_C_genotypes}{A binary (0/1) single-row matrix of genotypes for the
paired baseline isolates: named rows correspond to isolates: named columns
correspond to alleles; and matrix elements are set to 1 if the relevant allele
is called in a given isolate, and zero otherwise}

\item{ref_C_MOI}{MOI for the paired baseline isolate}

\item{error_matrix}{A row stochastic matrix (with row and column names
corresponding to alleles), where element (i,j) yields the probability
that an allele called as i in a baseline isolate matches an allele called
as j in the recurrent isolate}

\item{omega_vals}{Vector of values for omega, governing the locus-wise per-clone
probability of detection (applied to the paired baseline and recurrent isolates only)}
}
\value{
A data frame for the locus-wise likelihood (lhood) of the observed
genotypes for the recurrent isolate, given the number of newly-inoculated (n_I)
and recrudescent (n_C) clones comprising the recurrent isolate; and the per-clone
probability of detection for the paired baseline/recurrent isolates (omega)
}
\description{
This function evaluates the locus-wise likelihood of observations for the
recurrent isolate
}
