% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_likelihood.R
\name{locus_likelihood_given_detected_clones}
\alias{locus_likelihood_given_detected_clones}
\title{Locuswise likelihood conditional on detected clones}
\usage{
locus_likelihood_given_detected_clones(
  recurrent_alleles,
  recurrent_MOI,
  ref_I_genotypes,
  ref_I_MOI,
  ref_C_genotypes,
  ref_C_MOI,
  error_matrix
)
}
\arguments{
\item{recurrent_alleles}{Vector of alleles (strings) observed in the recurrent isolate}

\item{recurrent_MOI}{MOI of the recurrent isolate}

\item{ref_I_genotypes}{A binary (0/1) matrix of genotypes for baseline isolates
over which population I allele frequencies are derived: named rows correspond to isolates;
named columns correspond to alleles; and matrix elements are set to 1 if the
relevant allele is called in a given isolate, and zero otherwise}

\item{ref_I_MOI}{A named vector of MOIs for baseline isolates over which
population I allele frequencies are derived}

\item{ref_C_genotypes}{A binary (0/1) single-row matrix of genotypes for the
paired baseline isolates: named rows correspond to isolates: named columns
correspond to alleles; and matrix elements are set to 1 if the relevant allele
is called in a given isolate, and zero otherwise}

\item{ref_C_MOI}{MOI for the paired baseline isolate}

\item{error_matrix}{A row stochastic matrix (with row and column names
corresponding to alleles), where element (i,j) yields the probability
that an allele called as i in a baseline isolate matches an allele called
as j in the recurrent isolate}
}
\value{
A data frame for the conditional locus-wise likelihood (p) of the observed
genotypes for the recurrent isolate, given the number of newly-inoculated (n_I)
and recrudescent (n_C) clones comprising the recurrent isolate; and the number
of succesfully genotyped clones in the paired baseline (genotyped_C) and
recurrent (genotyped_recurrent) isolates respectively
}
\description{
This function evaluates the locus-wise likelihood of observations for the recurrent
isolate conditional on the number of successfully genotyped clones in the paired
baseline and recurrent isolates
}
