% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_moments.R
\name{locus_powerset_moments}
\alias{locus_powerset_moments}
\title{Calculate moments for sample allele counts}
\usage{
locus_powerset_moments(
  recurrent_alleles,
  recurrent_MOI,
  baseline_genotypes,
  baseline_MOI,
  error_matrix
)
}
\arguments{
\item{recurrent_alleles}{Vector of alleles (strings); allele counts will be
computed for the power set of these alleles}

\item{recurrent_MOI}{Maximum moment ordinal}

\item{baseline_genotypes}{A binary (0/1) matrix of genotypes for isolates
over which allele counts will be summed: named rows correspond to isolates;
named columns correspond to alleles; and matrix elements are set to 1 if the
relevant allele is called in a given isolate, and zero otherwise}

\item{baseline_MOI}{A named vector of MOIs for isolates over which allele
counts will be summed}

\item{error_matrix}{A row stochastic matrix (with row and column names
corresponding to alleles), where element (i,j) yields the probability
that an allele called as i in a baseline isolate matches an allele called
as j in the recurrent isolate}
}
\value{
A matrix of moments of allele counts summed over baseline isolates for
subsets of recurrent_alleles: columns correspond to allelic subsets (governed by
inc_exclusion_mask); rows correspond to the moment ordinal (running from zero
to recurrent_MOI)
}
\description{
This function calculates moments of allele couns for the power set of a given
set.of alleles, summed over one or more baseline isolates
}
