% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_moments.R
\name{sample_af_transform}
\alias{sample_af_transform}
\title{Convert moments of sample allele counts to moments of sample allele frequencies}
\usage{
sample_af_transform(pop_moments_not_in_subset, recurrent_MOI, n_clones)
}
\arguments{
\item{pop_moments_not_in_subset}{A matrix of moments of sample allele counts:
columns correspond to allelic subsets, rows correspond to the moment ordinal
(starting from zero)}

\item{recurrent_MOI}{Maximum moment ordinal}

\item{n_clones}{Number of clones over which allele counts have been computed}
}
\value{
A matrix of moments of moments of sample allele frequencies
}
\description{
Convert moments of sample allele counts to moments of sample allele frequencies
}
