% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_matrices.R
\name{uniform_error_matrix}
\alias{uniform_error_matrix}
\title{Uniform error model}
\usage{
uniform_error_matrix(marker_set, EPSILON)
}
\arguments{
\item{marker_set}{A named marker-wise list of vectors, listing all possible
alleles (strings) at each marker}

\item{EPSILON}{Genotyping error rate}
}
\value{
A named marker-wise list of row stochastic matrices (with row and column names
corresponding to alleles), where element (i,j) yields the probability
that an allele called as i in a baseline isolate matches an allele called
as j in the recurrent isolate
}
\description{
Adapted to nominal marker data: genotyping error rate EPSILON; distributed
uniformly over all possible alleles at the marker
}
