% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_posterior.R
\name{prior_mixture}
\alias{prior_mixture}
\title{Prior disrtibution}
\usage{
prior_mixture(n_C, recurrent_MOI)
}
\arguments{
\item{n_C}{Number of recrudescent clones}

\item{recurrent_MOI}{MOI of recurrent isolate}
}
\description{
This function computes the prior distribution of recrudescent (vs
newly-inoculated) clones within a recurrent isolate
}
